[IMP] im_livechat, *: introduce PublicLivechatWindow/renderMessages
*: website_livechat

Task-2964493
