+++ addons/im_livechat/static/src/models/public_livechat_window.js
@@ -47,6 +47,30 @@ registerModel({
                 .addClass('text-center fst-italic bg-200')
                 .val(disableText);
         },
+        renderMessages() {
+            const shouldScroll = !this.isFolded && this.publicLivechatView.widget.isAtBottom();
+            this.widget.render();
+            if (shouldScroll) {
+                this.publicLivechatView.widget.scrollToBottom();
+            }
+            const self = this;
+
+            this.widget.$('.o_thread_message:last .o_livechat_chatbot_options li').each(function () {
+                $(this).on('click', self.messaging.publicLivechatGlobal.livechatButtonView.widget._onChatbotOptionClicked.bind(self.messaging.publicLivechatGlobal.livechatButtonView.widget));
+            });
+
+            this.widget.$('.o_livechat_chatbot_main_restart').on('click',
+                this.messaging.publicLivechatGlobal.livechatButtonView.onChatbotRestartScript
+            );
+
+            if (this.messaging.publicLivechatGlobal.messages.length !== 0) {
+                const lastMessage = this.messaging.publicLivechatGlobal.lastMessage;
+                const stepAnswers = lastMessage.widget.getChatbotStepAnswers();
+                if (stepAnswers && stepAnswers.length !== 0 && !lastMessage.widget.getChatbotStepAnswerId()) {
+                    this.disableInput(this.env._t("Select an option above"));
+                }
+            }
+        },
         /**
          * @private
          * @returns {string}
